// src/modules/coupon/enums/coupon-error-code.enum.ts
export enum CouponErrorCode {
  VOUCHER_NOT_FOUND = 'VOUCHER_NOT_FOUND',
  VOUCHER_EXPIRED = 'VOUCHER_EXPIRED',
  VOUCHER_NOT_STARTED = 'VOUCHER_NOT_STARTED',
  VOUCHER_INACTIVE = 'VOUCHER_INACTIVE',
  USAGE_LIMIT_EXCEEDED = 'USAGE_LIMIT_EXCEEDED',
  USER_USAGE_LIMIT_EXCEEDED = 'USER_USAGE_LIMIT_EXCEEDED',
  MINIMUM_ORDER_NOT_MET = 'MINIMUM_ORDER_NOT_MET',
  PRODUCTS_NOT_APPLICABLE = 'PRODUCTS_NOT_APPLICABLE',
  FIRST_TIME_CUSTOMER_ONLY = 'FIRST_TIME_CUSTOMER_ONLY',
  CART_EMPTY = 'CART_EMPTY',
  ALREADY_COLLECTED = 'ALREADY_COLLECTED',
  NOT_STACKABLE = 'NOT_STACKABLE',
  MODIFICATION_LOCKED = 'MODIFICATION_LOCKED',
}